requirements.txt
flask
pandas
torch
plotly
scikit-learn
python-dateutil
